+++ lib/msal-browser/test/interaction_client/SilentRefreshClient.spec.ts
@@ -20,8 +20,6 @@ import {
     AuthenticationScheme,
     RefreshTokenClient,
     CommonSilentFlowRequest,
-    NetworkManager,
-    RefreshTokenEntity,
     AccountEntity,
     CredentialType,
 } from "@azure/msal-common";
@@ -32,6 +30,7 @@ import {
 } from "../../src/error/BrowserAuthError.js";
 import { SilentRefreshClient } from "../../src/interaction_client/SilentRefreshClient.js";
 import { BrowserCacheManager } from "../../src/cache/BrowserCacheManager.js";
+import { FetchClient } from "../../src/network/FetchClient.js";
 
 const testIdTokenClaims: TokenClaims = {
     ver: "2.0",
@@ -228,8 +227,8 @@ describe("SilentRefreshClient", () => {
                     "getRefreshToken"
                 ).mockReturnValue(rtEntity);
                 jest.spyOn(
-                    NetworkManager.prototype,
-                    "sendPostRequest"
+                    FetchClient.prototype,
+                    "sendPostRequestAsync"
                 ).mockResolvedValue(TEST_TOKEN_RESPONSE);
             });
