+++ lib/msal-common/src/exports-common.ts
@@ -91,9 +91,7 @@ export {
     NetworkRequestOptions,
     StubbedNetworkModule,
 } from "./network/INetworkModule.js";
-export {
-    NetworkResponse,
-} from "./network/NetworkResponse.js";
+export { NetworkResponse } from "./network/NetworkResponse.js";
 export { ThrottlingUtils } from "./network/ThrottlingUtils.js";
 export { RequestThumbprint } from "./network/RequestThumbprint.js";
 export { IUri } from "./url/IUri.js";
