+++ lib/msal-common/src/network/NetworkResponse.ts
