+++ lib/msal-common/test/client/AuthorizationCodeClient.spec.ts
@@ -3402,8 +3402,8 @@ describe("AuthorizationCodeClient unit tests", () => {
             ).mockResolvedValue(DEFAULT_OPENID_CONFIG_RESPONSE.body);
             jest.spyOn(
                 // @ts-ignore
-                client.networkManager,
-                "sendPostRequest"
+                client.networkClient,
+                "sendPostRequestAsync"
             ).mockResolvedValue(AUTHENTICATION_RESULT_WITH_HEADERS);
 
             if (!config.cryptoInterface) {
@@ -3451,6 +3451,7 @@ describe("AuthorizationCodeClient unit tests", () => {
                     refreshTokenSize:
                         AUTHENTICATION_RESULT_WITH_HEADERS.body.refresh_token
                             ?.length,
+                    requestId: "xMsRequestId",
                 },
                 RANDOM_TEST_GUID
             );
@@ -3482,8 +3483,8 @@ describe("AuthorizationCodeClient unit tests", () => {
             ).mockResolvedValue(DEFAULT_OPENID_CONFIG_RESPONSE.body);
             jest.spyOn(
                 // @ts-ignore
-                client.networkManager,
-                "sendPostRequest"
+                client.networkClient,
+                "sendPostRequestAsync"
             ).mockResolvedValue({ ...AUTHENTICATION_RESULT, headers: {} });
 
             if (!config.cryptoInterface) {
@@ -3530,6 +3531,7 @@ describe("AuthorizationCodeClient unit tests", () => {
                     httpVerToken: "",
                     refreshTokenSize:
                         AUTHENTICATION_RESULT.body.refresh_token?.length,
+                    requestId: "",
                 },
                 RANDOM_TEST_GUID
             );
